fix(tvos): share config validator with service layer
